+++ components/PricesPane.css
